did it,  max sends osc messages back and now we are able to read them and character is animated
character can now move base on the pose training that i do with posenet and flucoma max. i also scaled down the character and added actual movement. size down character and also framerate. will want to to add building blocking but i have the edge of the world blocking at the moment. need to do the actual training to see how it goes. also added changes in the server_osc to make sure that everytime i open it it destroys past sessions or anything that is listening on the post 8081.